FCBH-2781 Fix roles model and roleUser model
